Refactor DbFacade constructor to resolve database path on init
The _connection_string, used to find where to read or seed the database, is no longer hardcoded
It is now defined as a nullable string instead, as a placeholder.
In the constructor, the _connection_string is first assigned the value of CHIRPDBPATH,
which may be a temporary environment variable, or null if none is provided.

If CHIRPDBPATH is null, i.e. not defined when running the program, a string
is created from the user's local temp path and the name of the database to be created (chirp.db)
This string is then assigned as the value of _connection_string, and can be used
for database reading or seeding.

Co-authored-by: Anders <aing@itu.dk>