Retire QueryBuilder method from DBFacade.cs
NO PAGINATION MADE YET

After approaching the issue of pagination, and learning that SQLite
can be used to define a range and amount of rows to SELECT by using
the keywords LIMIT and OFFSET (more on these below), it quickly became
obvious that building queries with several conditions in mind (in this
case Cheeps by a certain author, amount of cheeps per page and offset
depending on page number) would not be practical, since it would require
several arguments passed to the QueryBuilder, many of which might be NULL
if a condition does not apply.

I still like the idea of a QueryBuilder making custom queries based on
conditions, but the amount of different parameters that would have to be
accounted for suggest not having a QueryBuilder would be simpler,
instead having each particular function have it's own query.

Though this will result in specific functions like GetCheeps()
and GetCheepsFrommAuthor() having similar queries, which feels
a bit redundant, it will also be easier to read and understand
what conditions each function imposes.

LIMIT and OFFSET keywords:

LIMIT <int count> defines how many matching rows (Cheeps) to
                  SELECT at most starting from the top of the
                  table (remember ORDER BY)

OFFSET <int offset> defines how many matching rows to skip from
                    the top before starting to SELECT. I intend
                    to use this to determine at which row to
                    begin SELECTing based on the page number

Note to self: try not to make commit descriptions overly verbose,
              also describe stuff after you implement it or use it