Make CheepRepository use LINQ for queries instead of SQLite strings
The CheepRepository now uses the ChirpDBContext to access a database (DBFacade is no longer used since EF Core handles the database
creation and connecting now), and can send LINQ queries to it. Currently the CheepRepository passes the necessary info for views
as a CheepViewModel record, as before, but it will soon be changed to use DTOs instead. Also, since the new database has not yet
been seeded using the provided initializer, it shows no Cheeps (to be added)