Refactor ChirpDBContext to map old Author fields to IdentityUser database column
Essentially, I wanted to avoid having to refactor a bunch of object initializers in the DBInitializer script, as well as methods
in our app. I have done this by modifying the AuthorId and Name from the Author class to point to the newly inherited
Id and UserName from the IdentityUser class, so the Author fields are still valid references, but point to the IdentityUser's
fields instead.

First, I have made getter and setter methods for AuthorId and Name to point to IdentityUser's fields instead. This mainly helps
with the database initialization, as the DBInitializer can still recognize AuthorId and Name, but set the values into Id and UserName.
Therefore, we did not have to refactor AuthorId to Id and Name to UserName a billion times in the initializer.
NOTE: We probably could have just done a 'Refactor all instances', but I like this solution better because it is cool B).

I ran into another issue when LINQ had to use the new fields. LINQ cannot inherently know that these point to the IdentityUser's
fields as it tries to access the database columns directly. Seeing how the columns created are now called Id and UserName instead
of AuthorId and Name, it crashed and said it could not find the Author columns.

So, to fix this, I added a method OnModelCreating to the ChirpDBContext, which overrides DbContext's own, and specified
that the properties Name and AuthorId should still exist, but be mapped to the columns with IdentityUser field names
instead. This way, when LINQ encounters Author.AuthorId or Author.Name, it will know to look into the IdentityUser's
columns. As it turns out, it is allowed to map several properties to a single column.