Add directive to enable partial pages in Chirp
Added the @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers directive in the top of _Layout.cshtml.
This directive is necessary to enable the <partial> tag, which helps specify which partial page to load
within the page the tag is used in. Default generated UI already has the directive imported, but we need
to add it manually since we are not using a generated UI.

The <partial> tag will allow us to define a header with the necessary buttons in each of our Chirp pages,
without having to add the header to each page individually, since we're defining it in the _Layout.cshtml.