Make the thing work
Do not ask me how. This whole thing took way too long, with way too much
manual refactoring and hacky solutions that should not work, but do.
Focus on the fact that it works, lest you wish to go insane trying to
understand it. I like the concept of scaffolding, but I think this
toolset has some problems. Either that or it is a problem on my
machine.