Sending cheeps now works
When a user is logged in they can now send cheeps through the form that will appear on both the public timeline and private
timeline. We have added a constraint on cheeps length on the client side. The code resides in the CheepBoxPartial.cshtml
where the constraint is put on the input text for that form. We have also added backend validation, that ensures
that a cheep is of correct length before it is being added. Note this is currently only when a user posts a cheep through
the public timeline. Currently onPost methods which are used for sending cheeps, is duplicated, and each of these two
identical methods reside in Public.cshtml and UserTimeline.cshtml. Both use the BindingProperty CheepText to create cheeps.

Co-authored-by: Anders <aing@itu.dk>
Co-authored-by: Bruno <brma@itu.dk>