Made use of CheepBoxModel class for model binding
We use the CheepBoxModel class to define properties that is used in pur input form. In CheepboxPartial.cshtml a CheepBoxModel is now used
in order to properly bind cheepboxtext to our input form. This ensures that when the html page for cheepboxPartial is rendered, it will use
the CheepBoxModel to validate and set constraints for our input form