Changed private and usertimelines html and pagemodels
The pagemodel for both private and public timeline will now use the CheepBoxModel when they are creating cheeps. This makes for a bit less code duplication
Also the html pages for private and public timeline will now render the CheepBoxPartial.html page themselves, instead of having the layout page rendering it.
This allows us to pass the proper cheepBoxModel to CheepBoxPartial when it needs to be rendered.