Util class for playwright tests
Main purpose for this class is to set up a localhost server, so we can run playwright tests against our application.
this includes starting a process that navigates to our chirp.web folder and runs dotnet run.
Main issue i have had with this is that our application takes to long to start, and when the test tries to use the process,
it fails since the project has not completed building. To resolve this issue i tried to make the test thread sleep for about 20 seconds
to ensure it did not run any test before our application was up and running - this worked but it is not super dynamic. I consulted with
Chat-gpt which helped make a solution that is more dynamic, where we wait for a specific log from dotnet that indicates that our server is ready,
and then we proceed. Logic for this starts at line 32, and i have added comments explaining in more detail what is happening.

Co-authored-by: Chat-GPT