UI tests using playwright
I have added three tests that utilizes playwright to test UI functionality. Each test will run an isolated browser,
that will try to perform actions against our application. After all tests has run, all processes are teared down.
I think the tests are pretty self-explanatory and i have added clarifyng comments, but one thing to note, is that
The tests for Log in are trying to log in using a 'testuser' which is a user i have added to the database