Modify Author class to never have Follow collection as null
On initializing the Author, the Follows collection of authors is automatically set to an empty list.
This is to avoid Follows ever being null, since it should at worst be an empty list when the current
user follows no one else. This way we avoid unnecessary warnings during build time.