Add method ReadFromFollows in CheepRepository
This new method, like many others in the CheepRepository, uses LINQ to find Cheeps from the current author,
and any other authors that they follow. This is done with a subquery, that extracts a list of all the followed
authors from the current author.

Then, the main query looks at each Cheep, checks if it's written by the current user, otherwise checks
whether the AuthorId of the current Cheep is included in the result of the subquery (a list of authors)