Added Delete feature for cheeps. Currently only available from public timeline
The Delete Button will only be available when a user is signed in, and it will only appear on a users own cheeps.
Since DTO's does not expose cheepid i could not find a way to match DTO from Website with Cheep in db with use of CheepID,
This would be the ideal approach i beleive.However i have used all three fields from DTO(author,timestamp,and message)
when querying the database to find the correct cheep to remove. Downside is that if a user has posted two identical cheeps at the same second
method will throw an exeption. Method that deletes cheeps is found in CheepRepository. Chat-GPT helped with ensuring correct timestamp conversion,
between DTO timestamp and Cheep datetime object.

Co-Authored-by: Chat-GPT