Fixed the Unit test and made timestamps formats are consistent across environments
I ran into an issue when testing the DeleteCheeps method in wsl(i had made the method and test in a powershell env). When pushed to
github it failed in our actions. Why? because of Datetime.ToString. I now fixed datetime format issue by specifying,
toString format when making DTO's. This should ensure a consistent format, when comparing Cheep.Timpestamp to a DTO.Message, across environments