Remove extra log out step
When pressing Logout or Forget me, the user is redirected straight
to the public timeline. At the moment this is done with GET requests

Although not the best/safest practice, it is a step forward. Planning
on redirecting to the Logout page so it says 'You have successfully logged out'
and then the user can go to the public timeline, or gets redirected after a few seconds