Scaffold pages related to email registration
I scaffolded the default pages for email confirmation and registration, added
some redirection back to the public timeline, and made the messages a bit friendlier and personalized.

Also, when a user registers their email, they are redirected to the ConfirmEmail page, where a message
'Thank you for confirming, you will be redirected shortly' appears, and then redirects to public timeline.