Fixed Playwright tests, which failed due to new UI elements
Also added one PLaywright tests that registers a user, and one that Forgets(deletes) that user. Note that previously we had a seperate test user
Stored in our db, that we use in our End to End test. That user has now been deleted, as we now make one dynamically, in the tests, and then we
delete that user at the end of the tests. An order to the tests have been added because of this, as we need to ensure that registering(creation)
of test user happens first, and Forget(deletion) happens last.