Fixed a bug where our application would break if an empty cheep was shared.
Now if a user tries to share an empty cheep an error message will appear instead of an exeption being thrown.
We had to alter the onPost methods for user and public timeline. If a modelstate is not valid which is the case when an empty cheep is posted,
we ensure that our CheepDTO list is populated, before we return the page. If not we will return to a page where error message is correctly shown, but no
cheeps are shown on the respective timeline